# Optimize get_node_tag_name()