--- a/src/element_handler/list.rs
+++ b/src/element_handler/list.rs
@@ -1,8 +1,15 @@
-use crate::{node_util::get_parent_node_tag_name, text_util::concat_strings, Element};
+use crate::{
+    node_util::{get_node_tag_name, get_parent_node},
+    text_util::concat_strings,
+    Element,
+};
 
 pub(super) fn list_handler(element: Element) -> Option<String> {
-    let parent_tag = get_parent_node_tag_name(&element.node);
-    if parent_tag.is_some_and(|tag| tag == "li") {
+    let parent = get_parent_node(&element.node);
+    let is_parent_li = parent
+        .map(|p| get_node_tag_name(&p).is_some_and(|tag| tag == "li"))
+        .unwrap_or(false);
+    if is_parent_li {
         Some(concat_strings!(
             "\n",
             element.content.trim_matches(|ch| ch == '\n'),