--- a/src/node_util.rs
+++ b/src/node_util.rs
@@ -2,15 +2,15 @@ use std::rc::Rc;
 
 use markup5ever_rcdom::{Node, NodeData};
 
-pub(crate) fn get_node_tag_name(node: &Rc<Node>) -> Option<String> {
+pub(crate) fn get_node_tag_name(node: &Rc<Node>) -> Option<&str> {
     match &node.data {
-        NodeData::Document => Some("html".to_string()),
-        NodeData::Element { name, .. } => Some(name.local.to_string()),
+        NodeData::Document => Some("html"),
+        NodeData::Element { name, .. } => Some(&name.local),
         _ => None,
     }
 }
 
-pub(crate) fn get_parent_node_tag_name(node: &Rc<Node>) -> Option<String> {
+pub(crate) fn get_parent_node(node: &Rc<Node>) -> Option<Rc<Node>> {
     let value = node.parent.take();
     let Some(parent) = value.as_ref() else {
         return None;
@@ -22,6 +22,5 @@ pub(crate) fn get_parent_node_tag_name(node: &Rc<Node>) -> Option<String> {
     };
     // Put the parent back
     node.parent.set(value);
-    let tag = get_node_tag_name(&parent);
-    tag
+    Some(parent)
 }