~23% faster after fixing `get_node_tag_name()`

```diff
- convert()               time:   [88.528 ms 89.545 ms 90.670 ms]
+ convert()               time:   [72.249 ms 72.795 ms 73.367 ms]
```